+++ src/proto/misc1.pro
@@ -83,15 +83,6 @@ char_u *concat_fnames(char_u *fname1, char_u *fname2, int sep);
 char_u *concat_str(char_u *str1, char_u *str2);
 void add_pathsep(char_u *p);
 char_u *FullName_save(char_u *fname, int force);
-pos_T *find_start_comment(int ind_maxcomment);
-void do_c_expr_indent(void);
-int cin_islabel(void);
-int cin_iscase(char_u *s, int strict);
-int cin_isscopedecl(char_u *s);
-void parse_cino(buf_T *buf);
-int get_c_indent(void);
-int get_expr_indent(void);
-int get_lisp_indent(void);
 void prepare_to_exit(void);
 void preserve_exit(void);
 int vim_fexists(char_u *fname);
